Never skip the performance tests.
The previous commit had the effect that they were run by default,
and only were skipped when the env variable was zero.
But that does not make much sense.

Also, even with a high test verbosity level (`bin/test -vvvvvv`),
sometimes the skip reason was shown and sometimes not.
The length of the reason seemed to matter at some point, but I could not get a grip on that.